refactor(pages): split pure helpers out of pages.rs into submodules
pages.rs was ~2950 lines. Extract three self-contained helper groups into
sibling modules under pages/: time_format (relative-time/freshness),
script_json (<script> JSON embedding), and search_text (plain-text/snippet/
highlight). Handlers and view structs stay in mod.rs; format_relative_time is
re-exported to keep handlers::pages::format_relative_time working for entries.rs.
