+++ src/handlers/pages/time_format.rs
@@ -1,0 +1,84 @@
+//! Relative-time and freshness formatting helpers shared by the SSR page
+//! handlers. Pure functions over `chrono` datetimes — no request state.
+
+pub fn format_relative_time_compact(dt: Option<chrono::DateTime<chrono::Utc>>) -> String {
+    let Some(dt) = dt else {
+        return "—".to_string();
+    };
+    let duration = chrono::Utc::now().signed_duration_since(dt);
+    let seconds = duration.num_seconds();
+    if seconds < 60 {
+        "now".to_string()
+    } else if seconds < 3600 {
+        format!("{}m", duration.num_minutes())
+    } else if seconds < 86400 {
+        format!("{}h", duration.num_hours())
+    } else if seconds < 2_592_000 {
+        format!("{}d", duration.num_days())
+    } else if seconds < 31_536_000 {
+        format!("{}mo", duration.num_days() / 30)
+    } else {
+        format!("{}y", duration.num_days() / 365)
+    }
+}
+
+/// Format a datetime as a human-readable relative time string.
+/// Returns (relative_text, iso_datetime_for_tooltip).
+pub fn format_relative_time(dt: Option<chrono::DateTime<chrono::Utc>>) -> (String, String) {
+    match dt {
+        None => ("Never".to_string(), String::new()),
+        Some(dt) => {
+            let now = chrono::Utc::now();
+            let duration = now.signed_duration_since(dt);
+            let seconds = duration.num_seconds();
+            let relative = if seconds < 60 {
+                "Just now".to_string()
+            } else if seconds < 3600 {
+                let mins = duration.num_minutes();
+                format!("{} minute{} ago", mins, if mins == 1 { "" } else { "s" })
+            } else if seconds < 86400 {
+                let hours = duration.num_hours();
+                format!("{} hour{} ago", hours, if hours == 1 { "" } else { "s" })
+            } else if seconds < 2_592_000 {
+                let days = duration.num_days();
+                format!("{} day{} ago", days, if days == 1 { "" } else { "s" })
+            } else if seconds < 31_536_000 {
+                let months = duration.num_days() / 30;
+                format!("{} month{} ago", months, if months == 1 { "" } else { "s" })
+            } else {
+                let years = duration.num_days() / 365;
+                format!("{} year{} ago", years, if years == 1 { "" } else { "s" })
+            };
+            (relative, dt.to_rfc3339())
+        }
+    }
+}
+
+/// Compute freshness CSS class and key from feed_updated_at and fetched_at.
+pub fn compute_freshness(
+    feed_updated_at: Option<chrono::DateTime<chrono::Utc>>,
+    fetched_at: Option<chrono::DateTime<chrono::Utc>>,
+) -> (String, String) {
+    let now = chrono::Utc::now();
+    match feed_updated_at {
+        Some(updated) => {
+            let days = (now - updated).num_days();
+            if days <= 30 {
+                (String::new(), "fresh".to_string())
+            } else if days <= 90 {
+                ("feed-freshness-warning".to_string(), "warning".to_string())
+            } else {
+                ("feed-freshness-stale".to_string(), "stale".to_string())
+            }
+        }
+        None => match fetched_at {
+            Some(fetched) if (now - fetched).num_days() <= 30 => {
+                ("muted".to_string(), "fresh".to_string())
+            }
+            Some(fetched) if (now - fetched).num_days() <= 90 => {
+                ("feed-freshness-warning".to_string(), "warning".to_string())
+            }
+            _ => ("feed-freshness-stale".to_string(), "stale".to_string()),
+        },
+    }
+}
